+++ src/server.js
@@ -1,33 +1,36 @@
 import fs from 'fs';
 import path from 'path';
+import urljoin from 'url-join';
 import polka from 'polka';
 import sirv from 'sirv';
+import compression from 'compression';
+import bodyParser from 'body-parser';
 
-const compression = require('compression');
-
-function parseBundleTree(file) {
+function parseBundleTree(file, public_static_path) {
     const bundle_tree = JSON.parse(fs.readFileSync(path.join(__dirname, file)));
 
     const dfs_cache = {};
 
     function dfs(file) {
-        if (!dfs_cache[file]) {
+        const file_public_path = urljoin(public_static_path, file);
+        if (!dfs_cache[file_public_path]) {
             const css = [], js = [];
             if (bundle_tree[file].imports) {
                 for (const f of bundle_tree[file].imports) {
+                    const f_public_path = urljoin(public_static_path, f);
                     if (!bundle_tree[f].isAsset) {
                         const { css: c, js: j } = dfs(f);
                         css.push(...c);
                         js.push(...j);
-                        js.push(f);
+                        js.push(f_public_path);
                     } else if (bundle_tree[f].isAsset && f.endsWith('.css')) {
-                        css.push(f);
+                        css.push(f_public_path);
                     }
                 }
             }
-            dfs_cache[file] = { css, js };
+            dfs_cache[file_public_path] = { css, js };
         }
-        return dfs_cache[file];
+        return dfs_cache[file_public_path];
     }
 
     const clientImports = {};
@@ -35,8 +38,9 @@ function parseBundleTree(file) {
     for (const file in bundle_tree) {
         if (bundle_tree[file].isEntry !== true) continue;
         const { css, js } = dfs(file);
-        clientImports[file] = { css: [...new Set(css)], js: [...new Set(js)] };
-        serverClientMap[path.basename(file).split('.')[0]] = file;
+        const file_public_path = urljoin(public_static_path, file);
+        clientImports[file_public_path] = { css: [...new Set(css)], js: [...new Set(js)] };
+        serverClientMap[path.basename(file).split('.')[0]] = file_public_path;
     }
 
     return { clientImports, serverClientMap };
@@ -66,7 +70,7 @@ function render(component_path, props_json, head, html) {
 </html>`;
 }
 
-function buildPage(component, props_json, { clientImports, serverClientMap }, public_static_path) {
+function buildPage(component, props_json, clientImports, serverClientMap) {
     let Component, props, head, html;
     try {
         Component = require(path.join(__dirname, 'server', component));
@@ -86,53 +90,62 @@ function buildPage(component, props_json, { clientImports, serverClientMap }, pu
         return { error: 'rendering module', code: err };
     }
 
-    const component_file = serverClientMap[component];
-    const component_path = `${path.join(public_static_path, component_file)}`;
-    clientImports[component_file].css.forEach(f =>
-        head += `\n\t<link rel='stylesheet' href='${path.join(public_static_path, f)}'>`);
-    clientImports[component_file].js.forEach(f =>
-        head += `\n\t<script type='module' src='${path.join(public_static_path, f)}'></script>`);
+    const component_path = serverClientMap[component];
+    clientImports[component_path].css.forEach(f =>
+        head += `\n\t<link rel='stylesheet' href='${f}'>`);
+    clientImports[component_path].js.forEach(f =>
+        head += `\n\t<script type='module' src='${f}'></script>`);
     head += `\n\t<script type='module' src='${component_path}'></script>`;
     const page = render(component_path, props_json, head, html);
 
     return { page };
 }
 
-const { clientImports, serverClientMap } = parseBundleTree('client-tree.json');
-
 const HOST = process.env.NODE_HOST || '0.0.0.0';
 const PORT = process.env.NODE_PORT || 3000;
+const PUBLIC_STATIC_PATH = 'STATIC_PATH';
+const DEV = 'DEV_SERVER';
+
+const { clientImports, serverClientMap } = parseBundleTree('client-tree.json', PUBLIC_STATIC_PATH);
 
 const server = polka();
-if (DEV_SERVER) {// eslint-disable-line
+if (DEV) {
     server.use(compression({ threshold: 0 }));
-    server.use('STATIC_PATH', sirv(path.join(__dirname, 'client'), {
+    let p = PUBLIC_STATIC_PATH.replace(/^(^\w+:|^)\/\/[\w.:]+/, '').replace(/[/]*$/, '');
+    server.use(p, sirv(path.join(__dirname, 'client'), {
+        setHeaders: res => {
+            res.setHeader('Access-Control-Allow-Origin', '*');
+        },
         maxAge: 31536000,
         immutable: true
     }));
 }
-server.get('/*', (req, res) => {
+const respond = (req, res) => {
+    let code, html;
     try {
-        const result = buildPage(req.params['*'] ? req.params['*'] : 'index',
-            req.query.props ? req.query.props : '{}',
-            { clientImports, serverClientMap }, 'STATIC_PATH');
+        const props = req.method === 'GET' ? (req.query.props ? req.query.props : '{}') : (req.body ? req.body : '{}');
+        const result = buildPage(req.params['*'] ? req.params['*'] : 'index', props, clientImports, serverClientMap);
         if (result.error) {
-            res.statusCode = 500;
-            res.end(`Error ${result.error}:\n${result.code}.`);
+            code = 500;
+            html = `<!DOCTYPE html><html lang="en"><body>Error ${result.error}:<br/>${result.code}.</body></html>`;
         } else {
-            const html = result.page;
-            res.writeHead(200, {
-                'Content-Type': 'text/html',
-                'Content-Length': html.length,
-            });
-            res.end(html);
+            code = 200;
+            html = result.page;
         }
     } catch (err) {
         console.error(err); // eslint-disable-line
-        res.statusCode = 500;
-        res.end(`Unknown error:\n${err}.`);
+        code = 500;
+        html = `<!DOCTYPE html><html lang="en"><body>Unknown error:<br/>${err}.</body></html>`;
     }
-})
+    res.writeHead(code, {
+        'Content-Type': 'text/html; charset=utf-8',
+        'Content-Length': html.length,
+    });
+    res.end(html);
+};
+server.use(bodyParser.text({ type: '*/json' }));
+server.get('/*', respond)
+    .post('/*', respond)
     .listen(PORT, HOST, err => {
         if (err) throw err;
         console.log(`Running on http://${HOST}:${PORT}`); // eslint-disable-line